requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.47.2
    # via
    #   -r requirements.in
    #   langchain-anthropic
anyio==4.8.0
    # via
    #   anthropic
    #   httpx
    #   openai
attrs==25.1.0
    # via aiohttp
boto3==1.37.1
    # via -r requirements.in
botocore==1.37.1
    # via
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via
    #   anthropic
    #   openai
faiss-cpu==1.10.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jiter==0.8.2
    # via
    #   anthropic
    #   openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.19
    # via
    #   -r requirements.in
    #   langchain-community
langchain-anthropic==0.3.8
    # via -r requirements.in
langchain-community==0.3.18
    # via -r requirements.in
langchain-core==0.3.40
    # via
    #   langchain
    #   langchain-anthropic
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
langchain-openai==0.3.7
    # via -r requirements.in
langchain-text-splitters==0.3.6
    # via langchain
langgraph==0.2.75
    # via -r requirements.in
langgraph-checkpoint==2.0.16
    # via langgraph
langgraph-sdk==0.1.53
    # via langgraph
langsmith==0.3.11
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
msgpack==1.1.0
    # via langgraph-checkpoint
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==2.2.3
    # via
    #   faiss-cpu
    #   langchain
    #   langchain-community
openai==1.64.0
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.15
    # via
    #   langgraph-sdk
    #   langsmith
packaging==24.2
    # via
    #   faiss-cpu
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pytest
pathspec==0.12.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   anthropic
    #   langchain
    #   langchain-anthropic
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.0
    # via langchain-community
pyperclip==1.9.0
    # via -r requirements.in
pytest==8.3.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
s3transfer==0.11.2
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
sqlalchemy==2.0.38
    # via
    #   langchain
    #   langchain-community
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.9.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anthropic
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.3.0
    # via
    #   botocore
    #   requests
yarl==1.18.3
    # via aiohttp
zstandard==0.23.0
    # via langsmith
